requirements.txt
PySide6
MetaTrader5
python-dotenv
